Allow removal of EasyLogging++ from the build
On some implementations, it is desired not to have
EasyLogging++ included in the build of Librealsense.
E.g. if EasyLogging++ is used in some other project
that is linked to Librealsense.

This is due to the global initialization macro that
EasyLogging++ rely on.

To disable, run cmake with:
cmake -DBUILD_EASYLOGGINGPP=false ../

Signed-off-by: Soren Friis <soren.friis@intel.com>