Modify decimation processing block to use linear scaling instead of exponention. (API change) Median filter to be used for all kernel sizes.
Currently the reference code for small kernels uses median of depth values, while for
kernels >3 it run mean of disparity values.
The inconsistency creates a gap to be addressed. Meanwhile a generic median of depths
will be employed.

Removing additional sanity boundaries in temporal filter.
Disparity filter to use a scaled baseline according to the reference code.

Change-Id: Ifbbb7e86fd5b99840ea4f3670df0fb30430c621e