[Node.js] Expose all native errors to Javascript world
Native errors would be exposed no matter recoverable or unrecoverable.
This is to be consistent with how the C++ layer deal with errors
especially to make sure users conscious of native errors even for APIs
without return values. Thus make it easy to target the cause of errors.
Cases are also changed to adpat this.